move modal indebted to its folder
